Move Fq and Fr trait implementations outside of arkworks when possible
