+++ benches/sqrt.rs
@@ -1,6 +1,6 @@
-use decaf377::Fq;
 use ark_ff::{Field, PrimeField, Zero};
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
+use decaf377::Fq;
 use decaf377::{SqrtRatioZeta, ZETA};
 use rand_chacha::ChaChaRng;
 use rand_core::{RngCore, SeedableRng};
