+++ src/fields/fq/arkworks.rs
@@ -7,12 +7,9 @@ use ark_serialize::{
 };
 use ark_std::{rand, str::FromStr, string::ToString, One, Zero};
 use core::convert::TryInto;
-use core::hash::Hash;
 use core::{
-    cmp::Ordering,
     fmt::{Display, Formatter},
     iter,
-    ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Sub, SubAssign},
 };
 
 impl PrimeField for Fq {
@@ -171,266 +168,6 @@ impl FftField for Fq {
     const LARGE_SUBGROUP_ROOT_OF_UNITY: Option<Self> = None;
 }
 
-impl From<u128> for Fq {
-    fn from(other: u128) -> Self {
-        let mut result = BigInt::default();
-        result.0[0] = ((other << 64) >> 64) as u64;
-        result.0[1] = (other >> 64) as u64;
-        Self::from_bigint(result).unwrap()
-    }
-}
-
-impl From<u64> for Fq {
-    fn from(other: u64) -> Self {
-        Self::from_bigint(BigInt::from(other)).unwrap()
-    }
-}
-
-impl From<u32> for Fq {
-    fn from(other: u32) -> Self {
-        Self::from_bigint(BigInt::from(other)).unwrap()
-    }
-}
-
-impl From<u16> for Fq {
-    fn from(other: u16) -> Self {
-        Self::from_bigint(BigInt::from(other)).unwrap()
-    }
-}
-
-impl From<u8> for Fq {
-    fn from(other: u8) -> Self {
-        Self::from_bigint(BigInt::from(other)).unwrap()
-    }
-}
-
-impl From<bool> for Fq {
-    fn from(other: bool) -> Self {
-        Self::from_bigint(BigInt::from(u64::from(other))).unwrap()
-    }
-}
-
-impl Neg for Fq {
-    type Output = Self;
-
-    #[inline]
-    #[must_use]
-    fn neg(self) -> Self {
-        let neg = self.neg();
-        neg
-    }
-}
-
-impl<'a> AddAssign<&'a Self> for Fq {
-    #[inline]
-    fn add_assign(&mut self, other: &Self) {
-        *self = self.add(other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::ops::AddAssign<Self> for Fq {
-    #[inline(always)]
-    fn add_assign(&mut self, other: Self) {
-        *self = self.add(&other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::ops::AddAssign<&'a mut Self> for Fq {
-    #[inline(always)]
-    fn add_assign(&mut self, other: &'a mut Self) {
-        *self = self.add(other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::ops::Add<Self> for Fq {
-    type Output = Self;
-
-    #[inline]
-    fn add(self, other: Self) -> Self {
-        self.add(&other)
-    }
-}
-
-impl<'a> Add<&'a Fq> for Fq {
-    type Output = Self;
-
-    #[inline]
-    fn add(self, other: &Self) -> Self {
-        self.add(other)
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::ops::Add<&'a mut Self> for Fq {
-    type Output = Self;
-
-    #[inline]
-    fn add(self, other: &'a mut Self) -> Self {
-        self.add(other)
-    }
-}
-
-impl<'a> SubAssign<&'a Self> for Fq {
-    #[inline]
-    fn sub_assign(&mut self, other: &Self) {
-        *self = self.sub(other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::ops::SubAssign<Self> for Fq {
-    #[inline(always)]
-    fn sub_assign(&mut self, other: Self) {
-        *self = self.sub(&other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::ops::SubAssign<&'a mut Self> for Fq {
-    #[inline(always)]
-    fn sub_assign(&mut self, other: &'a mut Self) {
-        *self = self.sub(other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::ops::Sub<Self> for Fq {
-    type Output = Self;
-
-    #[inline]
-    fn sub(self, other: Self) -> Self {
-        self.sub(&other)
-    }
-}
-
-impl<'a> Sub<&'a Fq> for Fq {
-    type Output = Self;
-
-    #[inline]
-    fn sub(self, other: &Self) -> Self {
-        self.sub(other)
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::ops::Sub<&'a mut Self> for Fq {
-    type Output = Self;
-
-    #[inline]
-    fn sub(self, other: &'a mut Self) -> Self {
-        self.sub(other)
-    }
-}
-
-impl<'a> MulAssign<&'a Self> for Fq {
-    fn mul_assign(&mut self, other: &Self) {
-        *self = self.mul(other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::ops::MulAssign<Self> for Fq {
-    #[inline(always)]
-    fn mul_assign(&mut self, other: Self) {
-        *self = self.mul(&other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::ops::MulAssign<&'a mut Self> for Fq {
-    #[inline(always)]
-    fn mul_assign(&mut self, other: &'a mut Self) {
-        *self = self.mul(other);
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::ops::Mul<Self> for Fq {
-    type Output = Self;
-
-    #[inline(always)]
-    fn mul(self, other: Self) -> Self {
-        self.mul(&other)
-    }
-}
-
-impl<'a> Mul<&'a Fq> for Fq {
-    type Output = Self;
-
-    #[inline]
-    fn mul(self, other: &Self) -> Self {
-        self.mul(other)
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::ops::Mul<&'a mut Self> for Fq {
-    type Output = Self;
-
-    #[inline(always)]
-    fn mul(self, other: &'a mut Self) -> Self {
-        self.mul(other)
-    }
-}
-
-impl<'a> DivAssign<&'a Self> for Fq {
-    #[inline(always)]
-    fn div_assign(&mut self, other: &Self) {
-        self.mul_assign(&other.inverse().unwrap());
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::ops::DivAssign<Self> for Fq {
-    #[inline(always)]
-    fn div_assign(&mut self, other: Self) {
-        self.div_assign(&other)
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::ops::DivAssign<&'a mut Self> for Fq {
-    #[inline(always)]
-    fn div_assign(&mut self, other: &'a mut Self) {
-        self.div_assign(&*other)
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::ops::Div<Self> for Fq {
-    type Output = Self;
-
-    #[inline(always)]
-    fn div(mut self, other: Self) -> Self {
-        self.div_assign(&other);
-        self
-    }
-}
-
-impl<'a> Div<&'a Fq> for Fq {
-    type Output = Self;
-
-    #[inline]
-    fn div(mut self, other: &Self) -> Self {
-        self.mul_assign(&other.inverse().unwrap());
-        self
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::ops::Div<&'a mut Self> for Fq {
-    type Output = Self;
-
-    #[inline(always)]
-    fn div(mut self, other: &'a mut Self) -> Self {
-        self.div_assign(&*other);
-        self
-    }
-}
-
 impl Zero for Fq {
     #[inline]
     fn zero() -> Self {
@@ -440,20 +177,6 @@ impl Zero for Fq {
     #[inline]
     fn is_zero(&self) -> bool {
         *self == Fq::ZERO
-    }
-}
-
-#[allow(unused_qualifications)]
-impl core::iter::Sum<Self> for Fq {
-    fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
-        iter.fold(Self::zero(), core::ops::Add::add)
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::iter::Sum<&'a Self> for Fq {
-    fn sum<I: Iterator<Item = &'a Self>>(iter: I) -> Self {
-        iter.fold(Self::zero(), core::ops::Add::add)
     }
 }
 
@@ -468,21 +191,6 @@ impl One for Fq {
         *self == Fq::ONE
     }
 }
-
-#[allow(unused_qualifications)]
-impl core::iter::Product<Self> for Fq {
-    fn product<I: Iterator<Item = Self>>(iter: I) -> Self {
-        iter.fold(Self::one(), core::ops::Mul::mul)
-    }
-}
-
-#[allow(unused_qualifications)]
-impl<'a> core::iter::Product<&'a Self> for Fq {
-    fn product<I: Iterator<Item = &'a Self>>(iter: I) -> Self {
-        iter.fold(Self::one(), Mul::mul)
-    }
-}
-
 impl CanonicalDeserializeWithFlags for Fq {
     fn deserialize_with_flags<R: ark_std::io::Read, F: Flags>(
         mut reader: R,
@@ -602,21 +310,6 @@ impl Display for Fq {
         write!(f, "{}", string.trim_start_matches('0'))
     }
 }
-
-impl Ord for Fq {
-    #[inline(always)]
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.into_bigint().cmp(&other.into_bigint())
-    }
-}
-
-impl PartialOrd for Fq {
-    #[inline(always)]
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
 impl FromStr for Fq {
     type Err = ();
 
@@ -651,25 +344,6 @@ impl From<BigInt<4>> for Fq {
     #[inline(always)]
     fn from(int: BigInt<4>) -> Self {
         Fq::from_le_bytes_mod_order(&int.to_bytes_le())
-    }
-}
-
-impl Hash for Fq {
-    fn hash<H: core::hash::Hasher>(&self, state: &mut H) {
-        state.write(&self.to_bytes_le())
-    }
-}
-
-impl Default for Fq {
-    fn default() -> Self {
-        Fq::zero()
-    }
-}
-
-impl core::fmt::Debug for Fq {
-    fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
-        let bytes = self.into_bigint().to_bytes_be();
-        write!(f, "Fq(0x{})", hex::encode(bytes))
     }
 }
